project-Master_aws_ljj/ssfe/requirements2.txt
black==23.3.0
boto3==1.26.111
botocore==1.29.111
click==8.1.3
Flask==2.2.3
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
